address pr feedback
- remove pre-increment usage
- add explicit flag to track consumption
- remove unused header
- minimize diff in test file
