+++ utils/s2n_map.h
@@ -25,6 +25,7 @@ struct s2n_map_iterator {
     const struct s2n_map *map;
     /* Index of the entry to be returned on the next `s2n_map_iterator_next()` call. */
     uint32_t current_index;
+    bool consumed;
 };
 
 struct s2n_map *s2n_map_new();
@@ -40,4 +41,3 @@ S2N_RESULT s2n_map_size(struct s2n_map *map, uint32_t *size);
 S2N_RESULT s2n_map_iterator_init(struct s2n_map_iterator *iter, const struct s2n_map *map);
 S2N_RESULT s2n_map_iterator_next(struct s2n_map_iterator *iter, struct s2n_blob *value);
 S2N_RESULT s2n_map_iterator_has_next(const struct s2n_map_iterator *iter, bool *has_next);
-S2N_CLEANUP_RESULT s2n_map_iterator_free(struct s2n_map_iterator *iter);
